requirements.txt
flask
faiss-cpu
numpy
pandas
sentence-transformers
requests
matplotlib
seaborn
